Add CLI module and forward __main__
